requirements.txt
azure-core>=1.29.5
azure-identity>=1.15.0
Flask>=3.0.0
Flask-Cors>=4.0.0
gunicorn>=21.2.0
openai>=1.6.1
python-dotenv>=1.0.0
requests>=2.31.0
